Add login page with navbar and multi-language support
Created a new login page for the application. This includes a navigation bar for better user experience. Also added support for multiple languages to enhance accessibility for users from different regions.

Create database and populate language tables

Set up a new database for the application and populated the language-related tables. This will be used to support the multi-language feature in the frontend.